Organized the project structure
