+++ src/test/java/tests/ClaimsStubDBTest.java
@@ -1,9 +1,9 @@
-package claims;
+package tests;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
-
+import classes.*;
 class ClaimsStubDBTest {
 	
 	Claims claim1 = new Claims();
