+++ src/test/java/tests/ClaimsTest.java
@@ -1,9 +1,8 @@
-package claims;
+package tests;
 
 import static org.junit.jupiter.api.Assertions.*;
-
 import org.junit.jupiter.api.Test;
-
+import classes.*;
 class ClaimsStubDBTest {
 	
 	Claims claim1 = new Claims(1, 100 , 110, 2, "Open", "01/08/24",
