+++ src/test/java/tests/CustomerTest.java
@@ -1,4 +1,4 @@
-package test.java;
+package tests;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
