+++ src/test/java/tests/VehicleStubDB.java
@@ -1,3 +1,4 @@
+package tests;
 import java.util.ArrayList;
 
 /*
