+++ src/test/java/tests/VehicleTest.java
@@ -1,3 +1,4 @@
+package tests;
 
 import static org.junit.jupiter.api.Assertions.*;
